requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncio==3.4.3
requests==2.31.0
psutil==5.9.8
pytz==2024.1
urllib3==2.2.1
six==1.16.0